+++ resources/js/Pages/Managements/Vehicles/Index.jsx
@@ -72,6 +72,9 @@ const Vehicle = ()=>{
             toast.warning('Por favor seleccione un registro para eliminar!');
             return;
         }
+        if (!confirm('¿Estás seguro de eliminar?')) {
+            return;
+        }
         axios.delete(route('vehicles.destroy',selectedRecord))
         .then(response => {
             toast.success(response.data.message);
